# PythonApi/openapi.json
{"openapi": "3.1.0", "info": {"title": "Todos API", "description": "Todos API", "version": "v1"}, "paths": {"/": {"get": {"summary": "Redirect To Swagger", "operationId": "redirect_to_swagger__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/Todos": {"get": {"tags": ["Todos"], "summary": "Get Todos", "operationId": "GetTodos", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/TodoItem"}, "type": "array", "title": "Response Gettodos"}}}}}}, "post": {"tags": ["Todos"], "summary": "Create Todo", "operationId": "CreateTodo", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateTodoCommand"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "integer", "title": "Response Createtodo"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/Todos/{id}": {"put": {"tags": ["Todos"], "summary": "Update Todo", "operationId": "UpdateTodo", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UpdateTodoCommand"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["Todos"], "summary": "Delete Todo", "operationId": "DeleteTodo", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"CreateTodoCommand": {"properties": {"title": {"type": "string", "title": "Title"}}, "type": "object", "required": ["title"], "title": "CreateTodoCommand"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "TodoItem": {"properties": {"id": {"type": "integer", "title": "Id"}, "title": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Title"}, "isComplete": {"type": "boolean", "title": "Iscomplete", "default": false}}, "type": "object", "required": ["id"], "title": "TodoItem"}, "UpdateTodoCommand": {"properties": {"title": {"type": "string", "title": "Title"}, "isComplete": {"type": "boolean", "title": "Iscomplete"}}, "type": "object", "required": ["title", "isComplete"], "title": "UpdateTodoCommand"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}